fixed bugs/checkin module complete

+++ src/main/java/com/booking/brownfield/dao/BookingDao.java
@@ -5,7 +5,7 @@ import org.springframework.stereotype.Repository;
 
 import com.booking.brownfield.entity.Booking;
 @Repository
-public interface BookingDao extends CrudRepository<Booking, Integer>{
+public interface BookingDao extends CrudRepository<Booking, Long>{
 	
 	public Booking findByBookingNo(long bookingNo);
